Bug 847388 - Using wl0.1 as DHCP and NAT interface on Nexus S
For some reason, the driver and/or netd are bringing up the network on
wl0.1 when we expect this to be on wlan0 interface. Let's then use wl0.1
for both handling the DHCP request and for the NAT. This will allow to
use WiFi Tethering properly. One consequence of this workaround is that
you need to either manually unload and reload the WiFi kernel driver or
reboot the device after using WiFi Tethering, otherwise the stack ends
in an unexpected state.